Get the projectiles moving and rendering
