+++ zelda_test/src/com/halfbit/entities/MoblinSpear.java
@@ -1,7 +1,9 @@
-package com.halfbit;
+package com.halfbit.entities;
 
 import java.awt.Point;
 
+import com.halfbit.Entity;
+import com.halfbit.Frame;
 import com.halfbit.Entity.Facing;
 import com.halfbit.sprites.StandingSprite;
 
@@ -12,6 +14,27 @@ public class MoblinSpear extends Entity {
 			new StandingSprite(new Frame("Resources/Enemies/Moblin/LAMobSpearE.png", 0, 0)),
 			new StandingSprite(new Frame("Resources/Enemies/Moblin/LAMobSpearW.png", 0, 0)),
 			new StandingSprite(new Frame("Resources/Enemies/Moblin/LAMobSpearS.png", 0, 0)),
-			position);
+			new Point(position), direction);
 	}
+
+	@Override
+	public Entity update(Point position) {
+		switch (this.facing) {
+		case NORTH:
+			this.position.y -= 5;
+			break;
+			
+		case SOUTH:
+			this.position.y += 5;
+			break;
+		case EAST:
+			this.position.x += 5;
+			
+		case WEST:
+			this.position.x -= 5;
+		}
+		return null;
+	}
+	
+	
 }
